token operation in a separate file
